+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -201,14 +201,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 - Replaced argument `id` with `by` in fields fetching a single entity
 - Must update GraphQL queries to use the new `filter`, `pagination` and `sort` field arguments
-- Renamed module "Schema for the Admin" to "Schema Expose Admin Data"
+- Renamed module "Schema for the Admin" to "Expose Admin Data in the Schema"
 - Renamed scalar type `AnyScalar` to `AnyBuiltInScalar`
 - Renamed interface type `Elemental` to `Node`
 - Renamed field `Root.option` to `Root.optionValue`
 - All `date` fields (such as `Post.date`, `Media.date` and `Comment.date`) and `modified` fields are now of type `DateTime` (before they had type `String`)
 - Must update `content(format:PLAIN_TEXT)` to `rawContent`
 - Must update the inputs for mutations
-- Merged the "admin" fields with the non-admin versions: instead of having fields `posts` and `unrestrainedPosts`, now there is only field `posts`, and its `filter` argument can also receive input `status` when `Schema Expose Admin Data` is enabled
+- Merged the "admin" fields with the non-admin versions: instead of having fields `posts` and `unrestrainedPosts`, now there is only field `posts`, and its `filter` argument can also receive input `status` when `Expose Admin Data in the Schema` is enabled
 - `User.email` is treated as "admin" field
 - Removed modules: Access Control, Cache Control, Public/Private Schema Mode, and Low-Level Persisted Query Editing
 - Settings for several modules must be set again